🧹 refactor: merge new form plugins into one module
- new way of making forms is less verbose!
