+++ src/forms/plugin/new_user.rs
@@ -1,7 +1,1 @@
 use super::*;
-
-impl Form {
-    pub fn new_user() -> Option<Self> {
-        None
-    }
-}
